// src/enums/route.enum.ts
export enum AppRouteEnums {
  HOME = '/',
  REGISTER_OFF_CHAIN = '/register-off-chain',
  TRADING = '/trading',
  TRADING_V2 = '/trading/v2',
  TRADING_V3 = '/trading/v3',
  TRADING_V4 = '/trading/v4',
  TRADING_V5 = '/trading/v5',
  TRADING_V6 = '/trading/v6',
  TRADING_V7 = '/trading/v7',
  EXPLORER = '/explorer',
  TRANSFER = '/transfer',
  CLAIM_REWARDS = '/claim-rewards',
  GET_GEMS_V1 = '/get-gems-v1',
  GET_GEMS_V2 = '/get-gems-v2',
  GENERATE_ETH_ACCOUNTS = '/generate-eth-accounts',
  CHECK_WALLET = '/check-wallet',
  CHECK_GEM = '/check-gem',
  CHECK_NFTS = '/check-nfts',
  TRADING_REWARD_GEMS = '/trading-reward-gems',
}